Struct tor_units::BoundedInt32

pub struct BoundedInt32<const LOWER: i32, const UPPER: i32> { /* private fields */ }

A 32-bit signed integer with a restricted range.

This type holds an i32 value such that LOWER <= value <= UPPER

Limitations

If you try to instantiate this type with LOWER > UPPER, you will get an uninhabitable type. It would be better if we could check that at compile time, and prevent such types from being named.

Implementations

impl<const LOWER: i32, const UPPER: i32> BoundedInt32<LOWER, UPPER>

pub const LOWER: i32 = LOWER

Lower bound

pub const UPPER: i32 = UPPER

Upper bound

pub fn get(&self) -> i32

Return the underlying i32 value.

This value will always be between Self::LOWER and Self::UPPER, inclusive.

pub fn saturating_new(val: i32) -> Self

If val is within range, return a new BoundedInt32 wrapping it; otherwise, clamp it to the upper or lower bound as appropriate.

pub fn checked_new(val: i32) -> Result<Self, Error>

If val is an acceptable value inside the range for this type, return a new BoundedInt32. Otherwise return an error.

pub fn saturating_from(val: i32) -> Self

Convert from the underlying type, clamping to the upper or lower bound if needed.

Panics

This function will panic if UPPER < LOWER.

pub fn saturating_from_str(s: &str) -> Result<Self, Error>

Convert from a string, clamping to the upper or lower bound if needed.

Limitations

If the input is a number that cannot be represented as an i32, then we return an error instead of clamping it.